+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Internal `Innmind\Signals\Handler::asAsync()`
+
+### Deprecated
+
+- `Innmind\Signals\Handler::async()`
+
 ## 5.1.0 - 2026-08-16
 
 ### Changed